other: rename files, export modules
