+++ packages/next-tinacms-markdown/src/index.tsx
@@ -16,18 +16,6 @@ limitations under the License.
 
 */
 
-import { usePlugins } from 'tinacms'
-import { useMarkdownForm, MarkdownFile, Options } from './useMarkdownForm'
-
-/**
- * Registers a Local Form with TinaCMS for editing a Markdown File.
- */
-export function useLocalMarkdownForm(
-  markdownFile: MarkdownFile,
-  options?: Options
-) {
-  const [values, form] = useMarkdownForm(markdownFile, options)
-  usePlugins(form)
-
-  return [values, form]
-}
+export * from './use-global-markdown-form'
+export * from './use-local-markdown-form'
+export * from './use-markdown-form'
